Add initial release candidate
This adds the first snapshot of the project that should be good for releasing, a release candidate. There's a set of tests, a `pytest` project, written and used for testing this library, but these haven't been integrated with the library project yet, so are omitted for now.

See `./README.md' for packaging, installation, usage etc.