Add procedure for determining token values for tokens of types that do not feature a `value` attribute
Tokenization as defined in the specification effectively maps code points in the input text stream, to certain token types where the "value" is implied -- a comma token, for instance, is created when the `,` is encountered in the input stream. We envision needing the inverse function which returns the code point given a token's type. This commit adds an implementation of such function.