Fix a critical issue with the `selector_list` attribute of `QualifedRule`
The "prelude" component of a qualified rule is a list of component values, not just tokens. Parsing it as a selector list, per the `parse_selector_list` API in the `selectors` module, demands a list of _tokens_ (currently), so a `tokens` call is necessary to "flatten" the list of component values. This adds the call, before normalization.