Fix a selector parsing bug
Parsing e.g. `:is(:is(foo))` (as a selector) does not consume the last `)`, which is obviously erroneous behaviour.

The bug stems from the body of the `parse_any_value` procedure in the `selectors` module which does not increment the counter number corresponding to parentheses when encountering a function token (which features these). With the wrong counter value, the first subsequently encountered `)` token (a `CloseParenToken`) would break the `any-value` parsing procedure before the latter token is consumed, as it would be considered "non-matching" (borrowing the term from the "Selectors" spec.).

This change fixes the issue by adding a `case` for `FunctionToken`, and supplements it all with an additional test, that was missing as well (meaning lack of coverage hid the issue all up until now, unfortunately).